+++ src/main/java/io/github/sefiraat/networks/network/barrel/BarrelType.java
@@ -3,6 +3,7 @@ package io.github.sefiraat.networks.network.barrel;
 public enum BarrelType {
     NETWORKS,
     INFINITY,
+    INFINITY_2,
     FLUFFY,
     TSINGSHAN_TECHNOLOGY,
     SIMPLE_STORAGE,