+++ src/main/java/io/github/sefiraat/networks/network/stackcaches/BarrelIdentity.java
@@ -2,10 +2,12 @@ package io.github.sefiraat.networks.network.stackcaches;
 
 import io.github.sefiraat.networks.network.barrel.BarrelCore;
 import io.github.sefiraat.networks.network.barrel.BarrelType;
+import io.github.sefiraat.networks.utils.StackUtils;
 import lombok.Getter;
 import lombok.Setter;
 import org.bukkit.Location;
 import org.bukkit.inventory.ItemStack;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.annotation.ParametersAreNonnullByDefault;
@@ -26,5 +28,10 @@ public abstract class BarrelIdentity extends ItemStackCache implements BarrelCor
         this.amount = amount;
         this.limit = limit;
         this.type = type;
+    }
+
+    /** Returns whether this storage identity can currently accept the supplied item. */
+    public boolean canAccept(@NotNull ItemStack incoming) {
+        return StackUtils.itemsMatch(this, incoming);
     }
 }